multilingual-assistant/requirements.txt
openai-whisper
torch
transformers
sounddevice
pydub
SpeechRecognition
gTTS
playsound
numpy